+++ tests/unit/core/test_sdk_api.py
@@ -16,21 +16,18 @@ class TestSdkApiClass(TestCase):
         common_props = {"host": "mock-url.com", "port": 443, "protocol": "https", "rejectUnauthorized": True}
         self.basic_props = {**common_props, "user": "Username", "password": "Password"}
         self.bearer_props = {**common_props, "tokenValue": "BearerToken"}
-        self.token_props = {
-            **common_props,
-            "tokenType": "MyToken",
-            "tokenValue": "TokenValue",
-        }
+        self.token_props = {**common_props, "tokenType": "MyToken", "tokenValue": "TokenValue"}
+        self.cert_props = {**common_props, "rejectUnauthorized": False, "certFile": "cert", "certKeyFile": "certKey"}
         self.default_url = "https://default-api.com/"
 
     def test_object_should_be_instance_of_class(self):
         """Created object should be instance of SdkApi class."""
         sdk_api = SdkApi(self.basic_props, self.default_url)
         self.assertIsInstance(sdk_api, SdkApi)
-    
-    @mock.patch('requests.Session.close') 
+
+    @mock.patch("requests.Session.close")
     def test_context_manager_closes_session(self, mock_close_request):
-    
+
         mock_close_request.return_value = mock.Mock(headers={"Content-Type": "application/json"}, status_code=200)
         with SdkApi(self.basic_props, self.default_url) as api:
             pass
@@ -47,13 +44,23 @@ class TestSdkApiClass(TestCase):
             self.assertIn("Host", mock_logger_error.call_args[0][0])
 
     @mock.patch("logging.Logger.error")
-    def test_session_no_authentication_logger(self, mock_logger_error: mock.MagicMock):
-        props = {"host": "test"}
+    def test_session_combined_cert_logger(self, mock_logger_error: mock.MagicMock):
+        props = {"host": "test", "certFile": "test"}
         try:
             sdk_api = SdkApi(props, self.default_url)
         except Exception:
             mock_logger_error.assert_called()
-            self.assertIn("Authentication", mock_logger_error.call_args[0][0])
+            self.assertIn("cert key", mock_logger_error.call_args[0][0])
+
+    def test_should_handle_none_auth(self):
+        props = {"host": "test"}
+        sdk_api = SdkApi(props, self.default_url)
+        self.assertEqual(sdk_api.session.password, None)
+
+    def test_should_handle_cert_auth(self):
+        props = self.cert_props
+        sdk_api = SdkApi(props, self.default_url)
+        self.assertEqual(sdk_api.session.cert, (self.cert_props["certFile"], self.cert_props["certKeyFile"]))
 
     def test_should_handle_basic_auth(self):
         """Created object should handle basic authentication."""